+++ app/Models/Sys/Vm.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models\Sys;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Vm extends Model
+{
+    use HasFactory;
+}