+++ routes/web.php
@@ -28,7 +28,7 @@ Route::domain("admin.diepxuan.com")->group(function () {
 
         Route::resources([
             'etc' => App\Http\Controllers\Sys\EnvController::class,
-            // 'posts' => PostController::class,
+            'vm' => App\Http\Controllers\Sys\VmController::class,
         ]);
     });
 });